move dicom sr logic to sr extension
