--- platform/core/src/DICOMSR/index.js
@@ -1,15 +1,0 @@
-import {
-  //retrieveMeasurements,
-  storeMeasurements,
-  downloadReport,
-} from './dataExchange';
-import isToolSupported from './utils/isToolSupported';
-
-const DICOMSR = {
-  //retrieveMeasurements,
-  storeMeasurements,
-  downloadReport,
-  isToolSupported,
-};
-
-export default DICOMSR;
